requirements.txt
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.82.0
FastAPI-SQLAlchemy==0.2.1
flower==1.2.0
greenlet==1.1.3
h11==0.13.0
humanize==4.3.0
idna==3.3
kombu==5.2.4
Mako==1.2.2
MarkupSafe==2.1.1
numpy==1.23.3
outcome==1.2.0
packaging==21.3
pandas==1.4.4
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pydantic==1.10.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
starlette==0.19.1
toml==0.10.2
tornado==6.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.2.0
